fix wirefree tests
